pyproject.toml
[project]
name = "catvdos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "gdown>=5.2.1",
    "pillow>=12.1.1",
    "streamlit>=1.55.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
]
